backend/tools/mcp_servers/requirements.txt
# MCP Servers - Python Dependencies
# Shared across all MCP tool servers

# Core MCP Framework
fastmcp>=0.1.0

# Async HTTP clients
httpx>=0.25.0

# LLM integration
openai>=1.0.0

# Code interpreter sandbox
docker>=7.0.0

# Testing (optional, only needed for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
